renamed all environments
