--- pettingzoo/classic/__init__.py
@@ -1,15 +1,0 @@
-from .chess import chess_env as chess_v0
-from .rps import rps as rps_v0
-from .rpsls import rpsls as rpsls_v0
-from .connect_four import connect_four as connect_four_v0
-from .tictactoe import tictactoe as tictactoe_v0
-from .rlcard_envs import leduc_holdem as leduc_holdem_v0
-from .rlcard_envs import mahjong as mahjong_v0
-from .rlcard_envs import texas_holdem as texas_holdem_v0
-from .rlcard_envs import texas_holdem_no_limit as texas_holdem_no_limit_v0
-from .rlcard_envs import uno as uno_v0
-from .rlcard_envs import dou_dizhu as dou_dizhu_v0
-from .rlcard_envs import gin_rummy as gin_rummy_v0
-from .go import go_env as go_v0
-from .hanabi import hanabi as hanabi_v0
-from .backgammon import backgammon_env as backgammon_v0
