+++ pettingzoo/classic/backgammon_v0.py
@@ -1,0 +1,1 @@
+from .backgammon.backgammon_env import env, raw_env
