+++ pettingzoo/classic/chess_v0.py
@@ -1,0 +1,1 @@
+from .chess.chess_env import env, raw_env
