+++ pettingzoo/classic/connect_four_v0.py
@@ -1,0 +1,1 @@
+from .connect_four.connect_four import env, raw_env
