+++ pettingzoo/classic/go_v0.py
@@ -1,0 +1,1 @@
+from .go.go_env import env, raw_env
