+++ pettingzoo/classic/hanabi_v0.py
@@ -1,0 +1,1 @@
+from .hanabi.hanabi import env, raw_env
