+++ pettingzoo/classic/mahjong_v0.py
@@ -1,0 +1,1 @@
+from .rlcard_envs.mahjong import env, raw_env
