+++ pettingzoo/classic/rps_v0.py
@@ -1,0 +1,1 @@
+from .rps.rps import env, raw_env
