+++ pettingzoo/classic/rpsls_v0.py
@@ -1,0 +1,1 @@
+from .rpsls.rpsls import env, raw_env
