+++ pettingzoo/classic/tictactoe_v0.py
@@ -1,0 +1,1 @@
+from .tictactoe.tictactoe import env, raw_env
