--- pettingzoo/gamma/__init__.py
@@ -1,5 +1,0 @@
-from .knights_archers_zombies import knights_archers_zombies as knights_archers_zombies_v0
-from .pistonball import pistonball as pistonball_v0
-from .cooperative_pong import cooperative_pong as cooperative_pong_v0
-from .prison import prison as prison_v0
-from .prospector import prospector as prospector_v0
