+++ pettingzoo/gamma/cooperative_pong_v0.py
@@ -1,0 +1,1 @@
+from .cooperative_pong.cooperative_pong import env, raw_env
