+++ pettingzoo/gamma/knights_archers_zombies_v0.py
@@ -1,0 +1,1 @@
+from .knights_archers_zombies.knights_archers_zombies import env, raw_env
