--- pettingzoo/magent/__init__.py
@@ -1,6 +1,0 @@
-from . import battle as battle_v0
-from . import gather as gather_v0
-from . import pursuit as adversarial_pursuit_v0
-from . import tiger as tiger_deer_v0
-from . import battlefield as battlefield_v0
-from . import combined_arms as combined_arms_v0
