+++ pettingzoo/magent/battlefield_v0.py
@@ -8,7 +8,7 @@ from pettingzoo.magent.render import Renderer
 from pettingzoo.utils import agent_selector, wrappers
 from .magent_env import magent_parallel_env, make_env
 from pettingzoo.utils._parallel_env import _parallel_env_wrapper
-from .battle import get_config
+from .battle_v0 import get_config
 
 
 def raw_env(map_size=80, seed=None):
