--- pettingzoo/mpe/__init__.py
@@ -1,9 +1,0 @@
-from . import simple_adversary as simple_adversary_v0
-from . import simple_crypto as simple_crypto_v0
-from . import simple_push as simple_push_v0
-from . import simple_reference as simple_reference_v0
-from . import simple_speaker_listener as simple_speaker_listener_v0
-from . import simple_spread as simple_spread_v0
-from . import simple_tag as simple_tag_v0
-from . import simple_world_comm as simple_world_comm_v0
-from . import simple as simple_v0
