+++ pettingzoo/mpe/simple_adversary_v0.py
