+++ pettingzoo/sisl/multiwalker_v0.py
@@ -1,0 +1,1 @@
+from .multiwalker.multiwalker import env, raw_env
