+++ pettingzoo/sisl/waterworld_v0.py
@@ -1,0 +1,1 @@
+from .waterworld.waterworld import env, raw_env
